refactor(track-page): extract TrackComments component
nate: 'loq relax or refactor — never golf.' the track page had been
relaxed five times today and passed 2000 lines; the comments world was
the natural cut now that it's sheet-contained. TrackComments.svelte owns
the state, loading (self-resetting on track.id change), compose, timed
seek-to-timestamp, edit/delete, the count bar, and the sheet — the page
drops from 2009 to 1211 lines and keeps only <TrackComments {track} />.
no behavior change; verified in-browser (bar count, sheet open, comment
render, timestamp chip, dismiss).
